requirements.txt
#https://askubuntu.com/questions/637014/gcc-error-trying-to-exec-cc1plus-execvp-no-such-file-or-directory

#sudo apt update && sudo apt install ffmpeg
#sudo apt-get install gcc
#sudo apt-get install g++
streamlit
streamlit-audiorecorder
gdown>=3.11
numpy>=1.19
scipy>=1.4
librosa>=0.8
vamp
lazycats
tensorflow>=2.6